Rename `intersection` module to `intersect`
The new name better matches the names of the other modules in
`algorithms`, as well as the name of the new `Intersect` trait.
