+++ crates/fj-kernel/src/algorithms/intersect/curve_edge.rs
