+++ crates/fj-kernel/src/algorithms/intersect/curve_face.rs
