+++ crates/fj-kernel/src/algorithms/intersect/face_face.rs
@@ -59,7 +59,7 @@ mod tests {
     use pretty_assertions::assert_eq;
 
     use crate::{
-        algorithms::intersection::CurveFaceIntersection,
+        algorithms::intersect::CurveFaceIntersection,
         objects::{Curve, Face, Surface},
     };
