+++ crates/fj-kernel/src/algorithms/intersect/line_segment.rs
