+++ crates/fj-kernel/src/algorithms/intersect/mod.rs
